add mood selection to firebase
everytime the user chooses a new mood it will be updated
